+++ vanguard/classification/dirichlet.py
@@ -49,7 +49,7 @@ class DirichletMulticlassClassification(Decorator):
         >>> test_x = np.array([0.05, 0.5, 0.95])
         >>> preds, probs = gp.classify_points(test_x)
         >>> preds
-        array([0, 1, 2])
+        array([0, 1, 2], dtype=int64)
     """
     def __init__(self, num_classes: int, **kwargs):
         """
